Refactor and move to /realtime app
